% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_raking.R
\name{allocate_pts}
\alias{allocate_pts}
\title{Distribuir amostra proporcionalmentem, levando em consideração o arredondamento.}
\usage{
allocate_pts(df, pts, min.pts)
}
\arguments{
\item{min.pts}{\emph{Inteiro} definindo o número mínimo de casos para um estrato}

\item{df.pts}{\emph{Dataframe} com a população de cada estrato. Nessa base devem haver pelo menos duas
colunas. Uma chamada strata, com os nomes dos estratos, e outra chamada pop, com a contagem populacional de cada estrato.}
}
\value{
Um \emph{Dataframe} com os dados originais, além do tamanho da amostra e algumas informações amostrais.
}
\description{
Essa função distribui a amostra proporcionalmente a população, controlando o arredondamento de forma
que o menor erro de arredondamento possível seja cometido. Ela também permite definir um número mínimo
de pontos por estrato.
}
\examples{

sample <- allocate_pts(df,pts,min.pts)
}
