% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{auto_cluster}
\alias{auto_cluster}
\title{Cluster Analysis combining Hierarquical and Kmeans cluster}
\usage{
auto_cluster(df = NULL, grps = 3:6, name = "kmeans", iter.max = 100)
}
\arguments{
\item{df}{A \emph{dataframe} containing the variables to be used in the analysis.}

\item{grps}{An \emph{array} or \emph{number} with the number of groups that should be created.}

\item{name}{A \emph{string} with the name of the variables that will be created.}

\item{iter.max}{A \emph{number} indicating the maximum number of iterations for the Kmeans cluster.}
}
\value{
A list with three components:
\itemize{
 \item \strong{grps}\emph{(dataframe)}: with the variables identifying the cluster each observation
 belongs too.
 \item \strong{summary}\emph{(dataframe)}: summary of the number of observations per cluster.
 \item \strong{details}\emph{(dataframe)}: details for each iteration of the kmeans algorithm.
}
}
\description{
This function runs as Cluster Analysis. The first step is to run a Hierarquical Cluster,
and then use centroids as starting point for the Kmeans Cluster. This function outputs the
cluster ids for each line in the dataframe and a summary of the group sizes. Details of the
analysis for each step are:
\itemize{
 \item \strong{Hierarquical}: Uses Euclidean distance and Ward's method. See more details
  in \code{\link[stats]{hclust}}.
 \item \strong{Kmeans}: Uses default options. See more details in \code{\link[stats]{kmeans}}.
}
}
\examples{
df.cluster <- auto_cluster(df=df,grps=3:6,name='kmeans')

}
