% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{clipboard_get}
\alias{clipboard_get}
\title{Import dataframe from clipboard generated by EXCEL. Before running this function in R, select
the appropriate cell range in EXCEL and press <CTRL+C>.}
\usage{
clipboard_get()
}
\value{
df The dataframe imported from Excel.
}
\description{
Import dataframe from clipboard generated by EXCEL. Before running this function in R, select
the appropriate cell range in EXCEL and press <CTRL+C>.
}
\examples{
clipboard_get()


}
