% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{clipboard_send}
\alias{clipboard_send}
\title{Export dataframe to clipboard. After running this function in R, select a target cell in EXCEL
and press <CTRL+V>.}
\usage{
clipboard_send(df)
}
\arguments{
\item{df}{The dataframe to be exported to Excel.}
}
\description{
Export dataframe to clipboard. After running this function in R, select a target cell in EXCEL
and press <CTRL+V>.
}
\examples{
clipboard_send(df)

}
