% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{df_summary}
\alias{df_summary}
\title{Summarises all variables in a dataframe}
\usage{
df_summary(df, drop = FALSE)
}
\arguments{
\item{df}{The dataframe to be summarised.}

\item{drop}{If \code{droplevels(df)} should be run before summarizing the dataframe.
This can make the orginal dataframe and the one analyzied differ. We
recomend you run \code{droplevels(df)} before calling this function if needed.}
}
\value{
A \emph{(dataframe)} with the description of all variables.
}
\description{
This function calculates the following statistics for all variables in the dataframe:
\itemize{
 \item \strong{var}: Variable name
 \item \strong{name}: Variable label
 \item \strong{class}: Variable class
 \item \strong{valid}: Number of valid cases
 \item \strong{n.na}: Number of \code{NA}'s
 \item \strong{n.distinct}: Number of distinct values
}
}
\examples{
summary <- df_summary(df)

}
