% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{dummy_all}
\alias{dummy_all}
\title{Creates an indicador variable for each category of the selected variables}
\usage{
dummy_all(df = NULL, reg.exp = NULL, keep_all = TRUE)
}
\arguments{
\item{df}{The dataframe containing the variables to be recoded.}

\item{reg.exp}{A \emph{regular expression} identifying the variables that should be analysed.}

\item{keep_all}{If \code{TRUE}, then for every category a dummy variable will be created. If
\code{FALSE} then the reference category will be dropped.}
}
\value{
A \emph{(dataframe)} contaning only the dummy variables that were created.
}
\description{
This function creates an indicador variable for each category of the selected variables
that belong to classes \emph{factor} or \emph{character}. Variables are selected using a
regular expression. The user can choose to keep all categories or to drop the reference
category. The new indicathor variables are named with the category label/value - only
alpha-numerical characters are kept, and all accents are removed. Empty categories (i.e.
factor levels that are defined but not used) are dropped.
}
\examples{
df.dummies <- dummy_all(df,reg.exp='^P[0-9]+$',keep_all=TRUE)

}
