% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{factor_analysis}
\alias{factor_analysis}
\title{Factor Analysis choosing number of factors automatically}
\usage{
factor_analysis(df = NULL, n.fat = NULL, name = "fator", sep = ".",
  rotation = "varimax", scores = "regression", cut = 0.2)
}
\arguments{
\item{df}{A \emph{dataframe} containing the variables to be used in the analysis.}

\item{n.fat}{A \emph{number} with the number of factors to use. if \emph{NULL} then the
number of factors is calculated automatically.}

\item{name}{A \emph{string} with the name of the factor variables that will be created.}

\item{sep}{A \emph{string} with the character to be used separating name of the factor
from the number of the factor}

\item{rotation}{A \emph{string} identifying the type of rotation to be performed.}

\item{scores}{A \emph{string} identifying the type of scores to be extracted.}

\item{cut}{A \emph{number} indicating the minimum size of factor loadings that should be kept.
Loadings smaller then \emph{cut} will be dropped.}
}
\value{
A list with two components:
\itemize{
 \item \strong{factors}\emph{(dataframe)}: with the variables with the scores extrated from
  the factor analysis.
 \item \strong{loadings}\emph{(dataframe)}: with the extracted factor loadings.
 \item \strong{details}\emph{(list)}: with all of the information returned from factanal.
}
}
\description{
This function runs as Factor Analysis. The first step is to calculate the number of factors
automatically. Once the number of factor is choosen, the Factor Analysis is run, and both
the estimated factor variables and the loadings are calculated. Missing values aren't aloud. Details
of the analysis for each step are:
\itemize{
 \item \strong{# of Factors}: Calculates the number of eigenvalues in the correlation matrix that
 are larger then 1. The idea is that if the variance of a standardized variable is 1, then a factor
 should only be maintained if it's variance is at least one. See more details in \code{\link[base]{eigen}}.
 \item \strong{Factor Analysis}: The default options are \code{rotation='varimax'} and
 \code{scores='regression'}. See more details in \code{\link[stats]{factanal}}.
}
}
\examples{
df_factan <- factor_analysis(df=df,cut=0.2)

}
