% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{get_spss}
\alias{get_spss}
\title{Load spss data with labels converted to factors}
\usage{
get_spss(file)
}
\arguments{
\item{file}{A characater string with the location of the spss file to be loaded.}
}
\value{
A list with two components:
\itemize{
 \item \strong{sav}\emph{(dataframe)}: with the actual data
 \item \strong{vars}\emph{(dataframe)}: with the description of all variables
}
}
\description{
This function loads an spss .sav file, and creates a dataframe with the description,
type and labels of all variables in the data file. This function automatically applies
as_factor to the data, converting all variables with labels in spss to factors in R.
}
\details{
To convert labels to factors function \code{haven::as_factor(.,only_labelled = TRUE,ordered=TRUE)} is used.
For details, see \code{?as_factor}.

The dataframe with the description of all variables contains the following information:
\itemize{
 \item \strong{var}: Variable name
 \item \strong{name}: Variable label
 \item \strong{class}: Variable class
 \item \strong{valid}: Number of valid cases
 \item \strong{n.na}: Number of \code{NA}'s
 \item \strong{n.distinct}: Number of distinct values
}
}
\examples{
spss.data <- get_spss(file)$sav

spss <- get_spss(file)
spss.data <- spss$sav
spss.vars <- spss$vars
}
