% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_raking.R
\docType{data}
\name{pop}
\alias{pop}
\title{Dados da Ambima}
\format{An object of class \code{data.frame} with 967 rows and 6 columns.}
\usage{
pop
}
\description{
Base de dados contendo perfil de investidores do Brasil, filtrada para classe AB e idade entre 25 e 70
}
\details{
@format Um dataframe com 967 casos e 6 variáveis:
 \describe{
 \item{nquest}{id do respondente}
 \item{pesoe}{peso do respondente}
 \item{regiao}{região do país}
 \item{sexo_cota}{sexo do respondente}
 \item{idade_cota}{idade do respondente}
 \item{classe_cota}{classe social do respondente}
 }
 @source \url{http://www.anbima.com.br/pt_br/especial/raio-x-do-investidor-2018.htm}
}
\keyword{datasets}
