% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_raking.R
\docType{data}
\name{svy}
\alias{svy}
\title{Pesquisa online com investidores}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1032 rows and 5 columns.}
\usage{
svy
}
\description{
Base de dados contendo perfil dos respondentes de uma pesquisa online, filtrada para classe AB e idade entre 25 e 70
}
\details{
@format Um dataframe com 1032 casos e 5 variáveis:
 \describe{
 \item{numericalId}{id do respondente}
 \item{regiao}{região do país}
 \item{sexo_cota}{sexo do respondente}
 \item{idade_cota}{idade do respondente}
 \item{classe_cota}{classe social do respondente}
 }
 @source \url{http://www.anbima.com.br/pt_br/especial/raio-x-do-investidor-2018.htm}
}
\keyword{datasets}
