% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_raking.R
\docType{data}
\name{svy.vote}
\alias{svy.vote}
\title{Pesquisa eleitoral online}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 52368 rows and 14 columns.}
\usage{
svy.vote
}
\description{
Base de dados de pesquisa online sobre intenção de voto nos EUA, filtrada para idade maior ou igual a 18 anos.
}
\details{

}
\keyword{datasets}
