% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{tab_summary}
\alias{tab_summary}
\title{Summary table is created, automatically adding marginal}
\usage{
tab_summary(df = NULL, reg.exp_lin = NULL, reg.exp_col = NULL, wgt = NULL)
}
\arguments{
\item{df}{A \emph{dataframe} containing the variables to be used in the analysis.}

\item{reg.exp_lin}{A \emph{regular expression} identifying the variables that should be used
to create the \emph{lines} of the table.}

\item{reg.exp_col}{A \emph{regular expression} identifying the variables that should be used
to create the \emph{columns} of the table.}

\item{wgt}{A \emph{string} identifying the weight variable. If \emph{NULL} weights are ignored.}
}
\value{
A \emph{(dataframe)} with the summary table.
}
\description{
This function is still incomplete, for now it only does counts and column total. NEED TO UPDATE!
}
\examples{
df.tab <- tab_summary(df=df,reg.exp_lin='^P1[0-4]$',reg.exp_col='^P2[6-9]$',wgt=NULL)

}
