% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{update_xlsx}
\alias{update_xlsx}
\title{Update xlsx workbook with a dataframe. A new spreadsheet can be created or updated.}
\usage{
update_xlsx(df, file, sheet)
}
\arguments{
\item{df}{The dataframe to be added to the workbook.}

\item{file}{A characater string with the location of the xlsx file.}

\item{sheet}{A characater string with the name of the sheet to be created/updated.}
}
\value{
Boolen value indicating if updating of the file was successful
}
\description{
Update xlsx workbook with a dataframe. A new spreadsheet can be created or updated.
}
\examples{
update_xlsx(df,wb,sheet)
}
