% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{View}
\alias{View}
\title{Invoke a Data Viewer, but removes list-columns so that R doesn't crash.}
\usage{
View(df)
}
\arguments{
\item{df}{The dataframe to be visualized.}
}
\value{
Invisible \emph{NULL}. The functions puts up a window and returns immediately: the window can be closed via its controls or menus.
}
\description{
Invoke a Data Viewer, but removes list-columns so that R doesn't crash.
}
\examples{

#generate list-column dataframe
df <- tibble(id=1:10,tab=map(1:10,~iris))

#utils View function
View(df)

View_(df)
}
