% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{write_spss}
\alias{write_spss}
\title{Save spss data}
\usage{
write_spss(df, file, max = 255)
}
\arguments{
\item{df}{The dataframe to be exported to sav format.}

\item{file}{A characater string with the location of the spss file to be saved.}

\item{max}{Maximum size of character strings.}
}
\description{
This function saves an spss '.sav' file removing known erros. String with length 0 are tranformed to " ". Also,
all character string are truncated to max length of paramenter 'max' and [[:cntrl:]] string are removed.
}
\examples{
spss.data <- write_sav(df,file)
}
